Refactor notifications to share animation logic
