+++ Assets/Scripts/UI/In-game notifications/AnimatedNotification.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 11ba0735d31a4be3b53b8f51ba628d8a
