+++ Assets/Scripts/UI/In-game notifications/TraitNotification.cs
@@ -1,16 +1,13 @@
-using System;
-using System.Collections;
+using GameAssets.ScriptableObjects.Core;
 using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
-using GameAssets.ScriptableObjects.Core;
 
 namespace VisualNovel.UI.Notifications
 {
     using UI;
-    using Animations;
-    
-    public class TraitNotification : GlobalNotification
+
+    public class TraitNotification : AnimatedNotification
     {
         [Header("UI Setup")]
         [SerializeField] private Image traitImage;
@@ -18,18 +15,6 @@ namespace VisualNovel.UI.Notifications
         [SerializeField] private TMP_Text traitDescriptionTextField;
         [SerializeField] private ExtendedButton readButton;
 
-        [Header("Animators")] 
-        [SerializeField] private AnimatorSequence[] animatorSequences;
-
-        [Serializable]
-        public class AnimatorSequence
-        {
-            public UIAnimator animator;
-            public float delayInSec;
-        }
-
-        private Coroutine showNotificationRoutine;
-        
         public void Initialize(TraitSO targetTrait)
         {
             traitImage.sprite = targetTrait.traitIcon;
@@ -38,71 +23,5 @@ namespace VisualNovel.UI.Notifications
         }
 
         public override ExtendedButton ReadButton => readButton;
-        
-        public override void Show()
-        {
-            if (showNotificationRoutine != null)
-            {
-                StopCoroutine(showNotificationRoutine);
-                showNotificationRoutine = null;
-            }
-            showNotificationRoutine = StartCoroutine(ShowNotificationRoutine());
-        }
-
-        public override void Hide()
-        {
-            if (showNotificationRoutine != null)
-            {
-                StopCoroutine(showNotificationRoutine);
-                showNotificationRoutine = null;
-            }
-
-            // hiding everything at once
-            if (animatorSequences == null) return;
-            foreach (var seq in animatorSequences)
-            {
-                if (seq?.animator != null)
-                    seq.animator.Play("Hide");
-            }
-        }
-
-        private IEnumerator ShowNotificationRoutine()
-        {
-            if (animatorSequences == null || animatorSequences.Length == 0)
-                yield break;
-
-            // If ALL delays are zero -> play all at once.
-            bool allZeroDelay = true;
-            foreach (var seq in animatorSequences)
-            {
-                if (seq != null && seq.delayInSec > 0f)
-                {
-                    allZeroDelay = false;
-                    break;
-                }
-            }
-
-            if (allZeroDelay)
-            {
-                foreach (var seq in animatorSequences)
-                {
-                    if (seq?.animator != null)
-                        seq.animator.Play("Appear");
-                }
-                yield break;
-            }
-
-            // Otherwise: play one-by-one, respecting each item's delay (relative to the previous trigger).
-            foreach (var seq in animatorSequences)
-            {
-                if (seq == null || seq.animator == null)
-                    continue;
-
-                if (seq.delayInSec > 0f)
-                    yield return new WaitForSeconds(seq.delayInSec);
-
-                seq.animator.Play("Appear");
-            }
-        }
     }
 }
